+++ packages/use/src/useRequest/index.test.ts
@@ -22,6 +22,7 @@ describe('useRequest', () => {
     expect(Object.keys(data).sort()).toStrictEqual([
       'response', 'data', 'error',
       'isExecuted', 'isLoading', 'isFinished', 'isSuccess',
+      'successCount', 'clearSuccessCount',
       'execute',
       'onSuccess', 'onError', 'onFinally',
       'reactive',
@@ -30,6 +31,7 @@ describe('useRequest', () => {
     expect(Object.keys(data.reactive).sort()).toStrictEqual([
       'response', 'data', 'error',
       'isExecuted', 'isLoading', 'isFinished', 'isSuccess',
+      'successCount', 'clearSuccessCount',
       'execute',
       'onSuccess', 'onError', 'onFinally',
     ].sort());
@@ -44,6 +46,9 @@ describe('useRequest', () => {
       { key: 'isLoading', type: 'boolean' },
       { key: 'isFinished', type: 'boolean' },
       { key: 'isSuccess', type: 'boolean' },
+
+      { key: 'successCount', type: 'number' },
+      { key: 'clearSuccessCount', type: 'function' },
 
       { key: 'execute', type: 'function' },
       { key: 'onSuccess', type: 'function' },
@@ -63,6 +68,7 @@ describe('useRequest', () => {
     expect(data.onSuccess).toBe(data.reactive.onSuccess);
     expect(data.onError).toBe(data.reactive.onError);
     expect(data.onFinally).toBe(data.reactive.onFinally);
+    expect(data.clearSuccessCount).toBe(data.reactive.clearSuccessCount);
 
     await data.execute();
 
@@ -129,6 +135,7 @@ describe('useRequest', () => {
     expect(data.isLoading.value).toBe(false);
     expect(data.isFinished.value).toBe(false);
     expect(data.isSuccess.value).toBe(false);
+    expect(data.successCount.value).toBe(0);
     expect(successEventCountAndArgs).toStrictEqual([0]);
     expect(errorEventCountAndArgs).toStrictEqual([0]);
     expect(finallyEventCountAndArgs).toStrictEqual([0]);
@@ -144,6 +151,7 @@ describe('useRequest', () => {
     expect(data.isLoading.value).toBe(true);
     expect(data.isFinished.value).toBe(false);
     expect(data.isSuccess.value).toBe(false);
+    expect(data.successCount.value).toBe(0);
     expect(successEventCountAndArgs).toStrictEqual([0]);
     expect(errorEventCountAndArgs).toStrictEqual([0]);
     expect(finallyEventCountAndArgs).toStrictEqual([0]);
@@ -157,6 +165,7 @@ describe('useRequest', () => {
     expect(data.isLoading.value).toBe(false);
     expect(data.isFinished.value).toBe(true);
     expect(data.isSuccess.value).toBe(true);
+    expect(data.successCount.value).toBe(1);
     expect(successEventCountAndArgs).toStrictEqual([1, { data: 1293 }]);
     expect(errorEventCountAndArgs).toStrictEqual([0]);
     expect(finallyEventCountAndArgs).toStrictEqual([1]);
@@ -173,6 +182,7 @@ describe('useRequest', () => {
     expect(data.isLoading.value).toBe(true);
     expect(data.isFinished.value).toBe(false);
     expect(data.isSuccess.value).toBe(false);
+    expect(data.successCount.value).toBe(1);
     expect(successEventCountAndArgs).toStrictEqual([1, { data: 1293 }]);
     expect(errorEventCountAndArgs).toStrictEqual([0]);
     expect(finallyEventCountAndArgs).toStrictEqual([1]);
@@ -191,6 +201,7 @@ describe('useRequest', () => {
     expect(data.isLoading.value).toBe(false);
     expect(data.isFinished.value).toBe(true);
     expect(data.isSuccess.value).toBe(false);
+    expect(data.successCount.value).toBe(1);
     expect(successEventCountAndArgs).toStrictEqual([1, { data: 1293 }]);
     expect(errorEventCountAndArgs).toStrictEqual([1, new Error('???')]);
     expect(finallyEventCountAndArgs).toStrictEqual([2]);
@@ -237,6 +248,7 @@ describe('useRequest', () => {
     expect(data.isLoading.value).toBe(false);
     expect(data.isFinished.value).toBe(false);
     expect(data.isSuccess.value).toBe(false);
+    expect(data.successCount.value).toBe(0);
     expect(successEventCountAndArgs).toStrictEqual([0]);
     expect(errorEventCountAndArgs).toStrictEqual([0]);
     expect(finallyEventCountAndArgs).toStrictEqual([0]);
@@ -252,6 +264,7 @@ describe('useRequest', () => {
     expect(data.isLoading.value).toBe(true);
     expect(data.isFinished.value).toBe(false);
     expect(data.isSuccess.value).toBe(false);
+    expect(data.successCount.value).toBe(0);
     expect(successEventCountAndArgs).toStrictEqual([0]);
     expect(errorEventCountAndArgs).toStrictEqual([0]);
     expect(finallyEventCountAndArgs).toStrictEqual([0]);
@@ -270,6 +283,7 @@ describe('useRequest', () => {
     expect(data.isLoading.value).toBe(false);
     expect(data.isFinished.value).toBe(true);
     expect(data.isSuccess.value).toBe(false);
+    expect(data.successCount.value).toBe(0);
     expect(successEventCountAndArgs).toStrictEqual([0]);
     expect(errorEventCountAndArgs).toStrictEqual([1, new Error('???')]);
     expect(finallyEventCountAndArgs).toStrictEqual([1]);
@@ -286,6 +300,7 @@ describe('useRequest', () => {
     expect(data.isLoading.value).toBe(true);
     expect(data.isFinished.value).toBe(false);
     expect(data.isSuccess.value).toBe(false);
+    expect(data.successCount.value).toBe(0);
     expect(successEventCountAndArgs).toStrictEqual([0]);
     expect(errorEventCountAndArgs).toStrictEqual([1, new Error('???')]);
     expect(finallyEventCountAndArgs).toStrictEqual([1]);
@@ -299,6 +314,7 @@ describe('useRequest', () => {
     expect(data.isLoading.value).toBe(false);
     expect(data.isFinished.value).toBe(true);
     expect(data.isSuccess.value).toBe(true);
+    expect(data.successCount.value).toBe(1);
     expect(successEventCountAndArgs).toStrictEqual([1, { data: 1293 }]);
     expect(errorEventCountAndArgs).toStrictEqual([1, new Error('???')]);
     expect(finallyEventCountAndArgs).toStrictEqual([2]);
@@ -346,6 +362,7 @@ describe('useRequest', () => {
     expect(successIndex).toBe(1);
     expect(finallyIndex).toBe(1);
     expect(errorIndex).toBe(0);
+    expect(data.successCount.value).toBe(1);
 
     throwError = true;
 
@@ -354,6 +371,57 @@ describe('useRequest', () => {
     expect(successIndex).toBe(1);
     expect(finallyIndex).toBe(2);
     expect(errorIndex).toBe(1);
+    expect(data.successCount.value).toBe(1);
+  });
+
+  test('请求成功后, 请求成功次数会进行累计, 可以使用方法清除请求成功次数', async () => {
+    let throwError: boolean = false;
+
+    const data = useRequest(async () => {
+      if (throwError) throw new Error('???');
+      return 123;
+    });
+
+    expect(data.successCount.value).toBe(0);
+
+    await data.execute();
+
+    expect(data.successCount.value).toBe(1);
+
+    await data.execute();
+
+    expect(data.successCount.value).toBe(2);
+
+    throwError = true;
+    try {
+      await data.execute();
+    }
+    catch (error) {
+      expect(error).toStrictEqual(new Error('???'));
+    }
+
+    expect(data.successCount.value).toBe(2);
+
+    try {
+      await data.execute();
+    }
+    catch (error) {
+      expect(error).toStrictEqual(new Error('???'));
+    }
+
+    expect(data.successCount.value).toBe(2);
+
+    throwError = false;
+    await data.execute();
+
+    expect(data.successCount.value).toBe(3);
+
+    data.clearSuccessCount();
+    expect(data.successCount.value).toBe(0);
+
+    await data.execute();
+
+    expect(data.successCount.value).toBe(1);
   });
 
   test('支持传入 immediate: true 选项立即发起请求', async () => {
@@ -373,6 +441,7 @@ describe('useRequest', () => {
     expect(data.isLoading.value).toBe(true);
     expect(data.isFinished.value).toBe(false);
     expect(data.isSuccess.value).toBe(false);
+    expect(data.successCount.value).toBe(0);
 
     await delay(100);
 
@@ -383,6 +452,7 @@ describe('useRequest', () => {
     expect(data.isLoading.value).toBe(false);
     expect(data.isFinished.value).toBe(true);
     expect(data.isSuccess.value).toBe(true);
+    expect(data.successCount.value).toBe(1);
   });
 
   test('支持传入 initialData 选项定义初始数据, 发起请求时会重置 data 为传入的 initialData', async () => {
@@ -727,6 +797,7 @@ describe('useRequestReactive', () => {
     expect(Object.keys(data).sort()).toStrictEqual([
       'response', 'data', 'error',
       'isExecuted', 'isLoading', 'isFinished', 'isSuccess',
+      'successCount', 'clearSuccessCount',
       'execute',
       'onSuccess', 'onError', 'onFinally',
     ].sort());
@@ -741,6 +812,9 @@ describe('useRequestReactive', () => {
       { key: 'isLoading', type: 'boolean' },
       { key: 'isFinished', type: 'boolean' },
       { key: 'isSuccess', type: 'boolean' },
+
+      { key: 'successCount', type: 'number' },
+      { key: 'clearSuccessCount', type: 'function' },
 
       { key: 'execute', type: 'function' },
       { key: 'onSuccess', type: 'function' },